cleanup deprecated components
