--- webapp/src/app/components/comments-new/comments-new.component.scss
@@ -1,48 +1,0 @@
-.comments {
-
-  .list {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-
-    .comment {
-      display: flex;
-      flex-direction: column;
-      gap: 1rem;
-      border-top: 1px solid #d9d9d9;
-      padding: 1rem 0;
-
-      .header {
-        display: flex;
-        align-items: center;
-        gap: 1rem;
-
-        .name {
-          cursor: pointer;
-        }
-
-        .date {
-          color: hsl(0, 0%, 75%);
-        }
-      }
-
-      .message {
-        padding-left: 30px;
-      }
-    }
-  }
-
-  .edit {
-    display: flex;
-    flex-direction: column;
-    margin: 1rem 0;
-
-    .button-container {
-      margin: 1rem 0;
-
-      button {
-        float: right;
-      }
-    }
-  }
-}
