+++ webapp/src/app/components/comments/comments.component.scss
@@ -7,46 +7,27 @@
 
     .comment {
       display: flex;
+      flex-direction: column;
+      gap: 1rem;
+      border-top: 1px solid #d9d9d9;
+      padding: 1rem 0;
 
-      .user{
+      .header {
         display: flex;
-        flex-direction: column;
         align-items: center;
-        width: 10rem;
+        gap: 1rem;
 
         .name {
-          color: black;
-          text-align: center;
+          cursor: pointer;
+        }
+
+        .date {
+          color: hsl(0, 0%, 75%);
         }
       }
 
       .message {
-        display: flex;
-        flex-direction: column;
-        flex-grow: 1;
-
-        .body {
-          background-color: hsl(0, 0%, 93%);
-          border-radius: 7px;
-          padding: 1.5rem;
-          position: relative;
-        }
-
-        .body:after {
-          content: " ";
-          position: absolute;
-          left: -8px;
-          top: 8px;
-          border-top: 8px solid transparent;
-          border-right: 8px solid hsl(0, 0%, 93%);
-          border-left: none;
-          border-bottom: 8px solid transparent;
-        }
-
-        .date {
-          text-align: right;
-          color: hsl(0, 0%, 88%);
-        }
+        padding-left: 30px;
       }
     }
   }
