+++ webapp/src/app/components/comments/comments.component.ts
@@ -2,33 +2,33 @@ import { Component, OnInit, Input, EventEmitter, Output } from '@angular/core';
 import { User } from '../../model/user';
 import { Comment } from '../../model/comment';
 import { AuthenticationService } from '../../services/authentication.service';
+import { BaseComponent } from '../base/base.component';
+import { UsersStore } from '../../core/store/user.store';
 
 @Component({
   selector: 'evendemy-comments',
   templateUrl: './comments.component.html',
   styleUrls: ['./comments.component.scss']
 })
-export class CommentsComponent implements OnInit {
+export class CommentsComponent extends BaseComponent {
 
   @Input()
   comments: any[] = [];
 
-  @Input()
-  users: User[] = [];
-
   @Output()
   addComment = new EventEmitter<Comment>();
-
+  
+  users: User[] = [];
   commentbox = '';
 
-  constructor(private authService: AuthenticationService) { }
-
-  ngOnInit() {
+  constructor(private authService: AuthenticationService, usersStore: UsersStore) { 
+    super();
+    this.addSubscription(usersStore.users().subscribe(users => this.users=users));
   }
 
   getUser(username: string) {
     const res = this.users.find( user => user.username === username);
-    return res ? res : username;
+    return res ? res : undefined;
   }
 
   onAddComment() {
