--- webapp/src/app/components/meeting-list/meeting-list.component.html
@@ -1,64 +1,0 @@
-<div class="meeting-list">
-  <div class="create-thumbnail" *ngIf="canCreate">
-      <div class="create-title" *ngIf="meetingType === 'meeting'">
-        <i class="fa fa-calendar fa-lg"></i>
-        <div>
-            create a meeting
-        </div>
-      </div>
-      <div class="create-title" *ngIf="meetingType === 'idea'">
-          <i class="fa fa-flask fa-lg"></i>
-          <div>
-              create an idea
-          </div>
-      </div>
-      <button type="button" class="btn btn-default create-button" routerLink="/{{meetingType}}">create</button>
-  </div>
-  <div class="thumbnail" *ngFor="let meeting of meetings">
-    <div class="caption text-center thumbnail-container">
-
-      <!-- picture -->
-      <div class="meeting-image">
-        <div class="meeting-image-info">
-            <div class="evendemy-badge" *ngIf="isMeetingNew(meeting)" >new</div>
-            <div  *ngIf="isAttending(meeting)" class="evendemy-tooltip">
-                <div class="status-attending" *ngIf="isAttending(meeting)"></div>
-                <span class="evendemy-tooltip-text right attendee-additional-badge-label">you want to attend</span>
-            </div>
-            <div *ngIf="hasTookPart(meeting)" class="evendemy-tooltip">
-                <div class="status-confirmed" *ngIf="hasTookPart(meeting)"></div>
-                <span class="evendemy-tooltip-text right attendee-additional-badge-label">you took part</span>
-            </div>
-        </div>
-        <img [src]="getImage(meeting.mid)" onError="this.src='assets/no-image.png'"/>
-      </div>
-
-      <!-- title -->
-      <h3 class="meeting-title">{{meeting.title}}</h3>
-
-      <!--subtitle-->
-      <p class="meeting-shortdescription" *ngIf="meeting.shortDescription">{{meeting.shortDescription}}</p>
-      <p class="no-meeting-shortdescription" *ngIf="!meeting.shortDescription">- - -</p>
-
-      <!--tags-->
-      <div class="meeting-tags" *ngIf="meeting.tags?.length > 0">
-        <evendemy-tag *ngFor="let tag of meeting.tags" [name]="tag"></evendemy-tag>
-      </div>
-      <p class="no-meeting-tags" *ngIf="meeting.tags?.length <= 0 || !meeting.tags">no tags</p>
-
-      <!-- time info -->
-      <div class="meeting-date">
-        <i class="fa fa-calendar" aria-hidden="true" style="padding-right: 5px;"></i>
-        <div style="display: inline-block;" *ngIf="meeting.startTime && meeting.endTime">
-          <div style="display:inline;">{{meeting.startTime | date:'dd.MM.yyyy'}}, {{meeting.startTime | date:'hh:mm'}}-{{meeting.endTime | date:'hh:mm'}}</div>
-        </div>
-        <div class="will-be-announced-label" *ngIf="!(meeting.startTime && meeting.endTime)">will be announced</div>
-      </div>
-
-      <!-- more-button -->
-      <a type="button" *ngIf="!meeting.isIdea" class="btn btn-default meeting-more-button" routerLink="/meeting/{{meeting.mid}}">more</a>
-      <a type="button" *ngIf="meeting.isIdea" class="btn btn-default meeting-more-button" routerLink="/idea/{{meeting.mid}}">more</a>
-
-    </div>
-  </div>
-</div>
