--- webapp/src/app/components/meeting-list/meeting-list.component.scss
@@ -1,141 +1,0 @@
-@import "./../../../variables";
-
-.thumbnail{
-    margin: 5px;
-    width: 260px;
-    height: 380px;
-}
-
-.create-thumbnail{
-  @extend .thumbnail;
-  text-align: center;
-  border: solid #dadada 2px;
-  border-radius: 5px;
-}
-
-// .create-thumbnail:hover{
-//   background-color: #dadada;
-// }
-
-.create-title{
-  padding-top: 120px;
-  font-size: 18px;
-  color: grey;
-  .fa{
-    padding-bottom: 5px;
-  }
-}
-
-.create-button{
-  margin-top: 10px;
-}
-
-.meeting-image-info{
-  position: absolute;
-  width: 230px;
-  text-align: right;
-  padding: 2px 5px 0px 2px;
-}
-
-.evendemy-badge{
-  display: inline-block;
-}
-
-.status {
-  display: inline-block;
-  border-radius: 10px;
-  padding: 6px;
-  margin-top: 5px;
-  border-color: white;
-  border-width: 1px;
-  border-style: solid;
-}
-
-.status-attending{
-  @extend .status;
-  background-color: #ffe11c;
-}
-
-.status-confirmed{
-  @extend .status;
-  background-color: $evendemy-primary;
-}
-
-.meeting-image {
-  position: relative;
-}
-
-.meeting-image > img{
-  width: 230px;
-  height: 172px;
-}
-
-.meeting-list{
-  display: -webkit-flex; /* Safari */
-  display: flex;
-  -webkit-flex-wrap: wrap; /* Safari */
-  flex-wrap: wrap;
-  -webkit-flex-direction: row; /* Safari */
-  flex-direction: row;
-}
-
-@media(max-width:500px){
-  .meeting-list{
-    justify-content: center;
-  }
-}
-
-.thumbnail-container{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 100%;
-}
-
-.meeting-shortdescription{
-  width: 230px;
-  height: 35px;
-  margin-top: 5px;
-  overflow:hidden;
-  white-space:nowrap;
-  text-overflow: ellipsis;
-}
-
-.no-meeting-shortdescription{
-  @extend .meeting-shortdescription;
-  color: $evendemy-background-alternate;
-}
-
-
-.meeting-title{
-  padding-top: 15px;
-  padding-bottom: 10px;
-  width: 230px;
-  overflow:hidden;
-  white-space:nowrap;
-  text-overflow: ellipsis;
-}
-
-.meeting-date{
-  margin-bottom: 10px;
-}
-
-.meeting-tags{
-  width: 230px;
-  height: 22px;
-  margin-bottom: 5px;
-  overflow: hidden;
-}
-
-.no-meeting-tags{
-  @extend .meeting-tags;
-  color: $evendemy-background-alternate;
-}
-
-.meeting-more-button{
-  margin-top: auto;
-}
-
-.will-be-announced-label{
-  display: inline-block;
-}
