--- webapp/src/app/components/page/page.component.html
@@ -1,11 +1,0 @@
-<div class="evendemy-page">
-  <div class="row">
-    <div class="col-md-10 col-md-offset-1">
-        <div class="jumbotron">
-            <div class="jumbotron-contents">
-                <ng-content></ng-content>
-            </div>
-          </div>
-    </div>
-  </div>
-</div>
