--- webapp/src/app/components/page/page.component.scss
@@ -1,3 +1,0 @@
-.evendemy-page{
-  padding: 20px;
-}
