--- webapp/src/app/components/page/page.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'evendemy-page',
-  templateUrl: './page.component.html',
-  styleUrls: ['./page.component.scss']
-})
-export class PageComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
