--- webapp/src/app/components/summary-courses-events/summary-courses-events.component.html
@@ -1,14 +1,0 @@
-<div style="display: flex; justify-content: center; align-items: center;">
-  <div style="display: flex; flex-direction: column; align-items: center; padding: 0px 10px;">
-    <h3 style="margin-top:0px;" class="attended-meetings">
-      {{numberOfAttendedMeetings}}
-    </h3>
-    attended
-  </div>
-  <div style="display: flex; flex-direction: column; align-items: center; padding: 0px 10px;">
-    <h3 style="margin-top:0px;" class="created-meetings">
-      {{numberOfCreatedMeetings}}
-    </h3>
-    created
-  </div>
-</div>
