--- webapp/src/app/pages/deprecated/events-or-courses/events-or-courses.component.html
@@ -1,39 +1,0 @@
-<div class="event-or-course">
-  <!-- filter bar -->
-  <div class="tool-bar">
-    <div>
-      Type:
-      <select [(ngModel)]="type" class="tool-bar-select" (click)="onToolbarChange()">
-            <option value="all">All</option>
-            <option value="event">Events</option>
-            <option value="course">Courses</option>
-      </select>
-      Date:
-      <evendemy-checkbox *ngIf="isIdea" [label]="'not announced'" (checkboxChanged)="onShowNotAnnounced($event)" [value]="showNotAnnounced"></evendemy-checkbox>
-      <evendemy-checkbox [label]="'old'" (checkboxChanged)="onShowOld($event)" [value]="showOld"></evendemy-checkbox>
-      <evendemy-checkbox [label]="'new'" (checkboxChanged)="onShowNew($event)" [value]="showNew"></evendemy-checkbox>
-      Tags:
-      <tag-input [(ngModel)]="selectedTags" name="tag" [modelAsStrings]="true"
-        [onlyFromAutocomplete]="true"
-        [placeholder]="'+tags'"
-        [secondaryPlaceholder]="'tags to be shown'"
-        [maxItems]="5" (onAdd)="onToolbarChange()" (onRemove)="onToolbarChange()"
-        class="tag-input">
-        <tag-input-dropdown [autocompleteItems]="allTags" [showDropdownIfEmpty]="true">
-        </tag-input-dropdown>
-      </tag-input>
-    </div>
-  </div>
-
-  <!-- meetings -->
-  <div  class="event-or-course-meetings">
-    <div class="loading-label" *ngIf="loading">loading...</div>
-    <evendemy-meeting-list
-      [meetings]="meetings"
-      [attendedMeetingInformation]="attendedMeetings"
-      [canCreate]="true"
-      [meetingType]="isIdea ? 'idea': 'meeting'"
-      ></evendemy-meeting-list>
-  </div>
-
-</div>
