--- webapp/src/app/pages/deprecated/events-or-courses/events-or-courses.component.scss
@@ -1,33 +1,0 @@
-.event-or-course-meetings{
-    padding: 15px;
-}
-.filter-label{
-    padding-right: 15px;
-    padding-left: 10px;
-    display: inline;
-}
-
-.tag-input{
-  display: inline-block;
-  padding-left: 10px;
-}
-
-.tool-bar{
-  text-align: right;
-  padding-right: 10px;
-}
-
-.tool-bar-select{
-  padding: 5px;
-  border-radius: 2px;
-  margin-right: 5px;
-  border: 0px;
-}
-
-.loading-label{
-  text-align: center;
-  padding: 200px 0px 200px 0px;
-  color: #6B6B6B;
-  font-family: 'Open Sans','Helvetica Neue',Helvetica,Arial,sans-serif!important;
-}
-
