--- webapp/src/app/pages/deprecated/meeting/meeting.component.scss
@@ -1,83 +1,0 @@
-@import './../../../../variables';
-
-.meeting-row{
-  margin: 0px;
-}
-
-.meeting-card{
-  padding: 20px 0px;
-}
-
-.meeting-image > img {
-  width: 100%;
-}
-
-.meeting-description-not-editable{
-  padding: 0px;
-  margin-bottom: 20px;
-}
-
-.meeting-image-button{
-  margin-top: -80px;
-}
-
-.meeting-csv-link{
-  margin-top: 30px;
-  padding-right: 0px;
-  text-align: right;
-  cursor: pointer;
-}
-
-.meeting-header-label{
-  padding-left: 0px;
-}
-
-.meeting-header-actions{
-  text-align: right;
-  margin-top:30px;
-
-  > a {
-    padding: 2px;
-    color: $evendemy-primary;
-  }
-
-  .not-active-link{
-    color: #c9c9c9;
-  }
-}
-
-.meeting-participants-info{
-  text-align: center;
-  padding-top: 10px;
-  color: grey;
-}
-
-.meeting-tag{
-  border-width: 1px;
-  border-style: solid;
-  border-radius: 20px;
-  border-color: $evendemy-primary;
-  color: $evendemy-primary;
-  padding: 0px 10px 0px 10px;
-}
-
-.meeting-checkbox{
-  margin-left: 15px;
-}
-
-.alert{
-  margin-top: 5px;
-  margin-bottom: 5px;
-}
-
-@import '~pretty-checkbox/src/scss/variables';
-@import './../../../../variables';
-
-$pretty--colors:(
-  primary: $evendemy-primary,
-);
-
-@import '~pretty-checkbox/src/pretty-checkbox.scss';
-
-@import './../../../override-tags-styling';
-
