--- webapp/src/app/pages/deprecated/user-info/user-info.component.html
@@ -1,163 +1,0 @@
-<div style="margin-top: 120px;">
-  <evendemy-page>
-    <div style="margin-top: -120px;">
-      <div style="display:flex; justify-content: center;">
-        <div style="text-align: center;">
-          <div class="img-tools">
-            <evendemy-user-image width="200" height="200" [username]="username"></evendemy-user-image>
-            <div class="edit-buttons" *ngIf="editable">
-              <i class="fa fa-pencil fa-lg" data-toggle="modal" data-target="#evendemy-image-upload-dialog"></i>
-              <i class="fa fa-trash fa-lg" data-toggle="modal" data-target="#deleteImageDialog"></i>
-            </div>
-          </div>
-
-          <p class="meeting-user-fullname">{{user?.firstname}} {{user?.lastname}}</p>
-          <p *ngIf="user.additional_info && user.options.additional_info_visible">
-            {{user.additional_info.job_title}}
-          </p>
-        </div>
-        <div>
-
-        </div>
-      </div>
-    </div>
-
-
-    <div *ngIf="user?.options && editable">
-      <h3 style="padding-top:20px;">Settings</h3>
-
-      <div>
-        <div class="pretty p-switch p-fill p-primary">
-          <input type="checkbox" (change)="onUpdateSettings()" [(ngModel)]="user.options.additional_info_visible" />
-          <div class="state">
-            <label>Additional Information visible</label>
-          </div>
-        </div>
-      </div>
-
-      <div>
-        <div class="pretty p-switch p-fill p-primary">
-          <input type="checkbox" (change)="onUpdateSettings()" [(ngModel)]="user.options.summary_of_meetings_visible" />
-          <div class="state">
-            <label>Created and Attended Meetings visible</label>
-          </div>
-        </div>
-      </div>
-
-    </div>
-
-    <div *ngIf="user.additional_info && editable">
-      <h3 style="margin-top:20px;">Additional Information</h3>
-      <div class="row info-row">
-        <div class="col-md-2">Job Title</div>
-        <div class="col-md-6">
-          <input type="text" class="form-control" [(ngModel)]="user.additional_info.job_title" name="jobTitle" (blur)="saveAdditionalInfo()" />
-        </div>
-      </div>
-
-      <div class="row info-row">
-        <div class="col-md-2">Facebook</div>
-        <div class="col-md-6">
-          <input type="text" class="form-control" [(ngModel)]="user.additional_info.facebook_username" name="facebook_username" (blur)="saveAdditionalInfo()" />
-        </div>
-      </div>
-
-      <div class="row info-row">
-        <div class="col-md-2">Instagram</div>
-        <div class="col-md-6">
-          <input type="text" class="form-control" [(ngModel)]="user.additional_info.instagram_username" name="instagram_username"  (blur)="saveAdditionalInfo()"/>
-        </div>
-      </div>
-
-      <div class="row info-row">
-        <div class="col-md-2">Twitter</div>
-        <div class="col-md-6">
-          <input type="text" class="form-control" [(ngModel)]="user.additional_info.twitter_username" name="twitter_username" (blur)="saveAdditionalInfo()" />
-        </div>
-      </div>
-
-      <div class="row info-row">
-        <div class="col-md-2">Description</div>
-        <div class="col-md-6">
-          <evendemy-editor [value]="user.additional_info.description" [editable]="editable" [headlines]="false" (change)="editorChanged($event)"></evendemy-editor>
-        </div>
-      </div>
-
-    </div>
-
-    <div *ngIf="user.additional_info && !editable && user.options.additional_info_visible">
-
-      <div class="social-media-icons">
-        <a *ngIf="user.additional_info.facebook_username" target="_blank" href="https://facebook.com/{{user.additional_info.facebook_username}}">
-          <i class="fa fa-facebook-official fa-lg"></i>
-        </a>
-        <a *ngIf="user.additional_info.instagram_username" target="_blank" href="https://instagram.com/{{user.additional_info.instagram_username}}">
-          <i class="fa fa-instagram fa-lg"></i>
-        </a>
-        <a *ngIf="user.additional_info.twitter_username" target="_blank" href="https://twitter.com/{{user.additional_info.twitter_username}}">
-          <i class="fa fa-twitter fa-lg"></i>
-        </a>
-      </div>
-
-      <!--readonly view of description-->
-      <div *ngIf="user.additional_info.description">
-        <div class="border-line"></div>
-        <evendemy-editor [value]="user.additional_info.description" [editable]="false"></evendemy-editor>
-      </div>
-
-    </div>
-
-    <!--Summary of created and attended meetings-->
-    <div *ngIf="user.options.summary_of_meetings_visible || editable">
-      <div class="border-line"></div>
-      <evendemy-summary-courses-events [numberOfAttendedMeetings]="events.length + courses.length" [numberOfCreatedMeetings]="events_from_author.length + courses_from_author.length">
-      </evendemy-summary-courses-events>
-    </div>
-
-    <div class="no-summary-row" *ngIf="!editable && !user.options.additional_info_visible">
-      <i class="fa fa-low-vision fa-2x no-summary-icon"></i>
-      <p>The user decided to not show his/her additional information.</p>
-    </div>
-
-  </evendemy-page>
-  <div class="row">
-    <div class="col-md-10 col-md-offset-1" style="padding: 0px 28px;">
-
-      <div class="user-info-row no-summary-row" *ngIf="!editable && !user.options.summary_of_meetings_visible">
-        <i class="fa fa-low-vision fa-2x no-summary-icon"></i>
-        <p>The user decided to not show his/her meetings as a summary.</p>
-      </div>
-
-      <div *ngIf="editable || user.options.summary_of_meetings_visible">
-        <h3 style="padding-top:20px;">{{courses.length}} attended Courses</h3>
-        <p *ngIf="courses.length==0">Nothing here at the moment...</p>
-        <evendemy-meeting-list [meetings]="courses"></evendemy-meeting-list>
-      </div>
-
-      <div *ngIf="editable || user.options.summary_of_meetings_visible">
-        <h3 style="padding-top:20px;">{{events.length}} attended Events</h3>
-        <p *ngIf="events.length==0">Nothing here at the moment...</p>
-        <evendemy-meeting-list [meetings]="events"></evendemy-meeting-list>
-      </div>
-
-      <div *ngIf="editable || user.options.summary_of_meetings_visible">
-        <h3 style="padding-top:20px;">{{courses_from_author.length}} created Courses</h3>
-        <p *ngIf="courses_from_author.length==0">Nothing here at the moment...</p>
-        <evendemy-meeting-list [meetings]="courses_from_author"></evendemy-meeting-list>
-      </div>
-
-      <div *ngIf="editable || user.options.summary_of_meetings_visible">
-        <h3 style="padding-top:20px;">{{events_from_author.length}} created Events</h3>
-        <p *ngIf="events_from_author.length==0">Nothing here at the moment...</p>
-        <evendemy-meeting-list [meetings]="events_from_author"></evendemy-meeting-list>
-      </div>
-
-
-      <evendemy-image-upload-dialog [forMeeting]='false' (data)="uploadImage($event)"></evendemy-image-upload-dialog>
-    </div>
-  </div>
-</div>
-
-<evendemy-confirm-dialog title="Do you really want to delete your account image?" (confirm)="deleteImage()" dialogID="deleteImageDialog">
-  Are you surce that you want to delete your account image?
-</evendemy-confirm-dialog>
