--- webapp/src/app/pages/deprecated/user-info/user-info.component.scss
@@ -1,93 +1,0 @@
-@import './../../../../variables';
-
-.info-row{
-  padding: 10px;
-}
-
-.social-media-icons{
-  display:flex;
-  justify-content: center;
-  a{
-    padding: 5px;
-    color: black;
-  }
-}
-
-.border-line{
-  width: 100%;
-  margin: 20px 0px;
-  border-bottom: 1px solid $evendemy-background;
-}
-
-.img-tools .edit-buttons {
-  opacity: 0;
-  margin-top: -100px;
-  padding-bottom: 100px;
-}
-
-.img-tools .edit-buttons .fa {
-  padding: 5px;
-}
-
-.img-tools .edit-buttons .fa:hover {
-  padding: 5px;
-  background-color: $evendemy-primary;
-  border-radius: 50%;
-}
-
-.img-tools:hover .edit-buttons {
-  opacity: 1;
-}
-
-.meeting-user-fullname{
-  font-weight: bold;
-  margin-top: 10px;
-  font-size: 2em;
-}
-.meeting-label{
-  border-radius: 10px;
-  background: orange;
-  padding: 1px 5px 0px 5px;
-  text-align: center;
-  margin-top:5px;
-  font-size: 12px;
-}
-.meeting-label:hover{
-  background: #fcd30a;
-}
-.meeting-label a{
-  color: rgb(67, 74, 84);
-}
-
-.event-label{
-  border-radius: 10px;
-  background: rgb(179,185, 185);
-  padding: 1px 5px 0px 5px;
-  text-align: center;
-  margin-top:5px;
-  font-size: 12px;
-}
-.event-label:hover{
-  background: rgb(206, 207, 207);
-}
-.event-label a{
-  color: rgb(67, 74, 84);
-}
-
-.user-info-row{
-  margin: 0px -15px 0px -15px;
-  padding: 0px 15px 15px 15px;
-}
-
-.user-info-row:nth-child(even){
-  background-color:#dddddd;
-}
-
-.no-summary-row{
-  text-align: center;
-  padding: 40px;
-}
-
-.no-summary-icon{
-  margin: 10px;
-}
