+++ webapp/src/app/pages/meeting-list/meeting-list.component.scss
