+++ webapp/src/app/pages/meeting/meeting.component.scss
