+++ webapp/src/app/pages/user-info/user-info.component.scss
